# openapi_specs.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HelpseekerCreationRequestSchema": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email",
          "password"
        ],
        "title": "HelpseekerCreationRequestSchema",
        "type": "object"
      },
...
      "UserSchema": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "user_type": {
            "$ref": "#/components/schemas/UserType"
          }
        },
        "required": [
          "id",
          "user_type",
          "first_name",
          "last_name",
          "email",
          "is_verified"
        ],
        "title": "UserSchema",
        "type": "object"
      },
...
      "UserType": {
        "enum": [
          "help-seeker",
          "volunteer"
        ],
        "title": "UserType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VolunteerCreationRequestSchema": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/VolunteerProperties"
            },
            "title": "Properties",
            "type": "array"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email",
          "password",
          "properties"
        ],
        "title": "VolunteerCreationRequestSchema",
        "type": "object"
      },
      "VolunteerProperties": {
        "enum": [
          "has_car",
          "can_host",
          "can_work_physical",
          "has_tools"
        ],
        "title": "VolunteerProperties",
        "type": "string"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "user/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "match",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/user/me": {
      "delete": {
        "operationId": "delete_me_user_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "user"
        ]
      },
      "get": {
        "operationId": "get_me_user_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "user"
        ]
      }
    },
...
    "/user/signup/helpseeker": {
      "post": {
        "operationId": "create_helpseeker_user_user_signup_helpseeker_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HelpseekerCreationRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Helpseeker User",
        "tags": [
          "user"
        ]
      }
    },
    "/user/signup/volunteer": {
      "post": {
        "operationId": "create_volunteer_user_user_signup_volunteer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VolunteerCreationRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Volunteer User",
        "tags": [
          "user"
        ]
      }
    },
    "/user/verify/{verification_code}": {
      "put": {
        "operationId": "verify_user_user_verify__verification_code__put",
        "parameters": [
          {
            "in": "path",
            "name": "verification_code",
            "required": true,
            "schema": {
              "title": "Verification Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Verify User User Verify  Verification Code  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify User",
        "tags": [
          "user"
        ]
      }
    },
    "/user/{user_id}": {
      "get": {
        "operationId": "get_user_user__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get User",
        "tags": [
          "user"
        ]
      }
    }
  }
}